requirements.txt
Flask==2.2.5
pymongo==4.6.0
python-dotenv==1.0.1

flask-mail==0.10.0
werkzeug==3.1.3

flask_pymongo
flask_cors

wkhtmltopdf==0.12.3

pip install pdfkit
pip install reportlab
